Make it possible to re-use TypePrettyPrinter
