+++ pmd-java/src/main/java/net/sourceforge/pmd/lang/java/types/WildcardTypeImpl.java
@@ -12,6 +12,7 @@ import org.checkerframework.checker.nullness.qual.Nullable;
 import org.pcollections.PSet;
 
 import net.sourceforge.pmd.lang.java.symbols.SymbolicValue.SymAnnot;
+import net.sourceforge.pmd.lang.java.types.prettyprint.TypePrettyPrint;
 
 
 final class WildcardTypeImpl implements JWildcardType {
