+++ pmd-java/src/test/kotlin/net/sourceforge/pmd/lang/java/types/TypeCreationDsl.kt
@@ -12,6 +12,7 @@ import net.sourceforge.pmd.lang.java.symbols.SymbolicValue.SymAnnot
 import net.sourceforge.pmd.lang.java.symbols.internal.FakeSymAnnot
 import net.sourceforge.pmd.lang.java.symbols.testdata.ClassWithTypeAnnotationsInside
 import net.sourceforge.pmd.lang.java.types.TypeOps.isSameType
+import net.sourceforge.pmd.lang.java.types.prettyprint.TypePrettyPrint
 import kotlin.reflect.KClass
